12/04/2082

+++ LargestThree/app/src/main/java/com/example/largestthree/MainActivity.java
@@ -17,7 +17,7 @@ public class MainActivity extends AppCompatActivity {
     EditText num3;
     TextView large;
     Button btn;
-    int lrg;
+    double lrg;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
